requirements.txt
# Mini Music Player Dependencies
# Author: eddy

customtkinter>=5.2.0
python-vlc>=3.0.18
mutagen>=1.46.0
requests>=2.31.0
Pillow>=10.0.0

# Optional - System tray support
pystray>=0.19.0

# Optional - Global hotkeys (Windows)
keyboard>=0.13.5
